// extensions/arc/src/controller/generated/dusky/model/duskyObjectModelsBackupRetention.ts
/**
 * Dusky API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export class DuskyObjectModelsBackupRetention {
    'maximums'?: Array<string>;
    'minimums'?: Array<string>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "maximums",
            "baseName": "maximums",
            "type": "Array<string>"
        },
        {
            "name": "minimums",
            "baseName": "minimums",
            "type": "Array<string>"
        }    ];

    static getAttributeTypeMap() {
        return DuskyObjectModelsBackupRetention.attributeTypeMap;
    }
}